Module citadel_sdk::fs

Convenience functions for interacting with the remote encrypted virtual filesystem (RE-VFS)

Functions

delete

Deletes a file from the Remote Encrypted Virtual Filesystem

read

Reads a file from the Remote Encrypted Virtual Filesystem

read_with_security_level

Reads a file from the Remote Encrypted Virtual Filesystem with a custom transport security level

take

Takes a file from the Remote Encrypted Virtual Filesystem

take_with_security_level

Takes a file from the Remote Encrypted Virtual Filesystem with a custom security level.

write

Writes a file or BytesSource to the Remote Encrypted Virtual Filesystem

write_with_security_level

Writes a file or BytesSource to the Remote Encrypted Virtual Filesystem with a custom security level.