Module citadel_sdk::prefabs::client::broadcast

A kernel that assists in creating and/or connecting to a group

Structs

BroadcastKernel

A kernel that streamlines creating, connecting, and interacting with groups Each group has a single owner, and, each connecting peer must at least be registered to the owner alone. The owner thus serves as an “axis of consent”, where each member trusts the owner, and through this trust, transitivity of trust flows to all other future members that connect to the group.

BroadcastShared

Enums

GroupInitRequestType

Before running the BroadcastKernel, each peer must send this request to the protocol. One peer can create a group, allowing others to join the group.