Enum citadel_sdk::prelude::NodeResult

pub enum NodeResult {

    RegisterOkay(RegisterOkay),
    RegisterFailure(RegisterFailure),
    DeRegistration(DeRegistration),
    ConnectSuccess(ConnectSuccess),
    ConnectFail(ConnectFail),
    ReKeyResult(ReKeyResult),
    OutboundRequestRejected(OutboundRequestRejected),
    ObjectTransferHandle(ObjectTransferHandle),
    MailboxDelivery(MailboxDelivery),
    PeerEvent(PeerEvent),
    GroupChannelCreated(GroupChannelCreated),
    GroupEvent(GroupEvent),
    Disconnect(Disconnect),
    InternalServerError(InternalServerError),
    PeerChannelCreated(PeerChannelCreated),
    SessionList(SessionList),
    Shutdown,

}

This type is for relaying results between the lower-level server and the higher-level kernel TODO: Convert to enum structs

Variants

RegisterOkay(RegisterOkay)

Returns the CNAC which was created during the registration process

RegisterFailure(RegisterFailure)

The registration was a failure

DeRegistration(DeRegistration)

When de-registration occurs. Third is_personal, Fourth is true if success, false otherwise

ConnectSuccess(ConnectSuccess)

Connection succeeded for the cid self.0. bool is “is personal”

ConnectFail(ConnectFail)

The connection was a failure

ReKeyResult(ReKeyResult)

OutboundRequestRejected(OutboundRequestRejected)

The outbound request was rejected

ObjectTransferHandle(ObjectTransferHandle)

For file transfers. Implicated CID, Peer/Target CID, object ID

MailboxDelivery(MailboxDelivery)

Mailbox

PeerEvent(PeerEvent)

Peer result

GroupChannelCreated(GroupChannelCreated)

For denoting a channel was created

GroupEvent(GroupEvent)

for group-related events. Implicated cid, ticket, group info

Disconnect(Disconnect)

vt-cxn-type is optional, because it may have only been a provisional connection

InternalServerError(InternalServerError)

An internal error occurred

PeerChannelCreated(PeerChannelCreated)

A channel was created, with channel_id = ticket (same as post-connect ticket received)

SessionList(SessionList)

A list of running sessions

Shutdown

For shutdowns

Implementations

impl NodeResult

pub fn is_connect_success_type(&self) -> bool

pub fn ticket(&self) -> Option<Ticket>